+++ src/main/java/com/tteokguk/tteokguk/member/application/dto/response/AppOAuthLoginResponse.java
@@ -1,4 +1,4 @@
-package com.tteokguk.tteokguk.member.application.dto;
+package com.tteokguk.tteokguk.member.application.dto.response;
 
 public record AppOAuthLoginResponse(
 	Long id,
